+++ src/components/Input/Input.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Input from "../components/Input/Input";
+import Input from "./Input";
 
 const meta = {
   title: "Input",
